+++ core-ng/src/main/java/core/framework/internal/log/filter/ArrayLogParam.java
@@ -14,19 +14,6 @@ public class ArrayLogParam implements LogParam {
 
     @Override
     public void append(StringBuilder builder, Set<String> maskedFields, int maxParamLength) {
-        int maxLength = builder.length() + maxParamLength;
-        builder.append('[');
-        for (int i = 0; i < values.length; i++) {
-            String value = values[i];
-            if (i > 0) builder.append(", ");
-            builder.append(value);
-
-            if (builder.length() > maxLength) {
-                builder.setLength(maxLength);
-                builder.append("...(truncated)");
-                return;
-            }
-        }
-        builder.append(']');
+        LogParamHelper.append(builder, values, maxParamLength);
     }
 }